moved map to servers component and RF servers store
