+++ react-app/src/components/Servers/ServerTile.js
@@ -1,36 +1,22 @@
-import React, { useEffect } from "react";
-import { useSelector, useDispatch } from "react-redux";
+import React from "react";
 import { Link } from "react-router-dom";
-
-import { getUserServers } from "../../store/user_servers";
 
 import styles from "./ServerTile.module.css";
 // className={styles. }
 
-export default function ServerTile() {
-    const dispatch = useDispatch();
-
-    const servers = useSelector((state) => Object.values(state.user_servers));
-    const user_id = useSelector((state) => state.session.user.id);
-
-    useEffect(() => {
-        dispatch(getUserServers(user_id));
-    }, [dispatch, user_id]);
-
+export default function ServerTile({ server }) {
+    if (!server) {
+        return null;
+    }
     return (
-        <div>
-            {servers &&
-                servers.map((server) => (
-                    <div key={server.id} className={styles.iconWrapper}>
-                        <Link to={`/servers/${server.id}`}>
-                            <img
-                                className={styles.icon}
-                                src={server.icon}
-                                alt="Server-Icon"
-                            />
-                        </Link>
-                    </div>
-                ))}
+        <div className={styles.iconWrapper}>
+            <Link to={`/servers/${server.id}`}>
+                <img
+                    className={styles.icon}
+                    src={server.icon}
+                    alt="Server-Icon"
+                />
+            </Link>
         </div>
     );
 }
